requirements.txt
streamlit>=1.33.0
spacy>=3.7.0
torch>=2.1.0
transformers>=4.39.3
pdfplumber>=0.10.2
newspaper3k>=0.2.8
pandas>=1.5.3